add: OrigoStorageService. focusing on getting photodownloader to work with origo
