--- src/main/groovy/com/cloudcard/photoDownloader/integrations/PeopleService.groovy
@@ -1,9 +1,0 @@
-package com.cloudcard.photoDownloader.integrations
-
-import com.cloudcard.photoDownloader.Person
-
-interface PeopleService {
-
-    List<Person> getPeople()
-
-}
